feat(poincarejs): add a cleanUpAfterRun function to the poincarejs module
This function ensures the memory is properly freed after its callback.

This also allows to use an async instantiation of the module,
and get rid of most of the side-functions in /poincare-partial.js
We now let each user of the lib instantiate the module the way they want.
All the missing code will be moved and modified inside FoxGlove.

The poincare-partial.js was replced with extern-pre-js.ts